fix: pixel receives just intergers

+++ src/gl/pixel.cxx
@@ -3,5 +3,5 @@
 namespace pixel
 {
     Pixel::Pixel() : x(0), y(0) {}
-    Pixel::Pixel(double x, double y) : x(x), y(y) {}
+    Pixel::Pixel(int x, int y) : x(x), y(y) {}
 }
